Add patrolling
Very much not a finished thing. I need to implement event triggering to
allow jumping to other parts of the tree when e.g. an enemy suddenly
becomes sighted. This is kind of broken right now, where the ai will
go to the last known location of its enemy but gets stuck forever if it
sees the enemy while it's running because it cannot now start running
the attack tree.